refactor: PageResponse 클래스 패키지 이동
